Use naming convention for all files
This ensures that all files are using a naming convention.
