+++ cmd/check_fritz/check_downstream.go
@@ -1,0 +1,39 @@
+package main
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/mcktr/check_fritz/pkg/fritz"
+)
+
+func CheckDownstreamMax(hostname, port string, username string, password string) {
+	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/wancommonifconfig1", "WANCommonInterfaceConfig", "X_AVM-DE_GetOnlineMonitor")
+	fritz.AddSoapRequestVariable(&soapReq, fritz.NewSoapRequestVariable("NewSyncGroupIndex", "0"))
+
+	err := fritz.DoSoapRequest(&soapReq)
+
+	if HandleError(err) {
+		return
+	}
+
+	var resp = fritz.GetWANCommonInterfaceOnlineMonitorResponse{}
+
+	err = fritz.HandleSoapRequest(&soapReq, &resp)
+
+	if HandleError(err) {
+		return
+	}
+
+	downstream, err := strconv.Atoi(resp.Newmax_ds)
+
+	if HandleError(err) {
+		return
+	}
+
+	downstream = downstream * 8 / 1000000
+
+	fmt.Print("OK - Max Downstream: " + strconv.Itoa(downstream) + " Mbit/s \n")
+
+	GlobalReturnCode = exitOk
+}
