+++ cmd/check_fritz/main.go
@@ -4,9 +4,6 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"strconv"
-
-	"github.com/mcktr/check_fritz/pkg/fritz"
 )
 
 // program version
@@ -46,140 +43,6 @@ func CheckRequiredFlags(hostname string, port string, username string, password 
 	}
 
 	return true
-}
-
-func CheckConnectionStatus(hostname string, port string, username string, password string) {
-	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/wanpppconn1", "WANPPPConnection", "GetInfo")
-
-	err := fritz.DoSoapRequest(&soapReq)
-
-	if HandleError(err) {
-		return
-	}
-
-	var resp = fritz.GetWANPPPConnectionInfoResponse{}
-
-	err = fritz.HandleSoapRequest(&soapReq, &resp)
-
-	if HandleError(err) {
-		return
-	}
-
-	if resp.NewConnectionStatus == "Connected" {
-		fmt.Print("OK - Connection Status: " + resp.NewConnectionStatus + "\n")
-
-		GlobalReturnCode = exitOk
-	} else {
-		fmt.Print("CRITICAL - Connection Status: " + resp.NewConnectionStatus + "\n")
-
-		GlobalReturnCode = exitCritical
-	}
-}
-
-func CheckConnectionUptime(hostname, port string, username string, password string) {
-	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/wanpppconn1", "WANPPPConnection", "GetInfo")
-
-	err := fritz.DoSoapRequest(&soapReq)
-
-	if HandleError(err) {
-		return
-	}
-
-	var resp = fritz.GetWANPPPConnectionInfoResponse{}
-
-	err = fritz.HandleSoapRequest(&soapReq, &resp)
-
-	if HandleError(err) {
-		return
-	}
-
-	fmt.Print("OK - Connection Uptime: " + resp.NewUptime + "\n")
-
-	GlobalReturnCode = exitOk
-}
-
-func CheckDeviceUptime(hostname, port string, username string, password string) {
-	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/deviceinfo", "DeviceInfo", "GetInfo")
-
-	err := fritz.DoSoapRequest(&soapReq)
-
-	if HandleError(err) {
-		return
-	}
-
-	var resp = fritz.GetDeviceInfoResponse{}
-
-	err = fritz.HandleSoapRequest(&soapReq, &resp)
-
-	if HandleError(err) {
-		return
-	}
-
-	fmt.Print("OK - Device Uptime: " + resp.NewUpTime + "\n")
-
-	GlobalReturnCode = exitOk
-}
-
-func CheckDownstreamMax(hostname, port string, username string, password string) {
-	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/wancommonifconfig1", "WANCommonInterfaceConfig", "X_AVM-DE_GetOnlineMonitor")
-	fritz.AddSoapRequestVariable(&soapReq, fritz.NewSoapRequestVariable("NewSyncGroupIndex", "0"))
-
-	err := fritz.DoSoapRequest(&soapReq)
-
-	if HandleError(err) {
-		return
-	}
-
-	var resp = fritz.GetWANCommonInterfaceOnlineMonitorResponse{}
-
-	err = fritz.HandleSoapRequest(&soapReq, &resp)
-
-	if HandleError(err) {
-		return
-	}
-
-	downstream, err := strconv.Atoi(resp.Newmax_ds)
-
-	if HandleError(err) {
-		return
-	}
-
-	downstream = downstream * 8 / 1000000
-
-	fmt.Print("OK - Max Downstream: " + strconv.Itoa(downstream) + " Mbit/s \n")
-
-	GlobalReturnCode = exitOk
-}
-
-func CheckUpstreamMax(hostname, port string, username string, password string) {
-	soapReq := fritz.NewSoapRequest(username, password, hostname, port, "/upnp/control/wancommonifconfig1", "WANCommonInterfaceConfig", "X_AVM-DE_GetOnlineMonitor")
-	fritz.AddSoapRequestVariable(&soapReq, fritz.NewSoapRequestVariable("NewSyncGroupIndex", "0"))
-
-	err := fritz.DoSoapRequest(&soapReq)
-
-	if HandleError(err) {
-		return
-	}
-
-	var resp = fritz.GetWANCommonInterfaceOnlineMonitorResponse{}
-
-	err = fritz.HandleSoapRequest(&soapReq, &resp)
-
-	if HandleError(err) {
-		return
-	}
-
-	upstream, err := strconv.Atoi(resp.Newmax_us)
-
-	if HandleError(err) {
-		return
-	}
-
-	upstream = upstream * 8 / 1000000
-
-	fmt.Print("OK - Max Upstream: " + strconv.Itoa(upstream) + " Mbit/s \n")
-
-	GlobalReturnCode = exitOk
 }
 
 func HandleError(err error) bool {
