+++ pkg/fritz/fritz_response.go
